feat: test #1084
