+++ src/i18n/locales/en/translations.json
@@ -240,7 +240,7 @@
   "sub-account": "Sub Account",
   "asset-balance": "Asset Balance",
   "value": "Value",
-  "tokex_expire": "Token Expired. Please re-loginin.",
+  "token_expired": "Token Expired. Please re-loginin.",
   "email-exist": "This email already has administrator privileges",
   "state": "State",
   "home-announcement-1": "Statement on cryptocurrency regulation and compliance",
