+++ src/i18n/locales/zh_CN/translations.json
@@ -239,7 +239,7 @@
   "sub-account": "子账号",
   "asset-balance": "资产水位",
   "value": "数量",
-  "tokex_expire": "页面闲置过久，请重新登入",
+  "token_expired": "页面闲置过久，请重新登入",
   "email-exist": "此 Email 已有管理者权限",
   "state": "状态",
   "home-announcement-1": "有关全球加密货币监管及合规声明",
