+++ src/pages/index.js
@@ -1,6 +1,8 @@
 import React, { Suspense, lazy, useContext, useEffect, useState } from "react";
-import { Switch, Route } from "react-router-dom";
+import { Switch, Route, useHistory } from "react-router-dom";
+import { useTranslation } from "react-i18next";
 import StoreContext from "../store/store-context";
+import Dialog from "../components/Dialog";
 // import Markets from "../pages/markets";
 // import Profile from "./profile";
 // import Wallet from "./wallet";
@@ -23,13 +25,14 @@ const Admin = lazy(() => import("./admin"));
 const Index = () => {
   const storeCtx = useContext(StoreContext);
   const [isInit, setIsInit] = useState(null);
+  const history = useHistory();
+  const { t } = useTranslation();
 
   useEffect(() => {
     if (isInit === null) {
       setIsInit(false);
       storeCtx.init().then((_) => setIsInit(true));
     }
-
     // ++TODO never called
     return () => {
       // storeCtx.stop();
@@ -38,19 +41,35 @@ const Index = () => {
   }, [isInit, storeCtx]);
 
   return (
-    <Suspense fallback={<div></div>}>
-      <Switch>
-        <Route exact path="/">
-          <Home />
-        </Route>
-        <Route path="/markets">
-          <Exchange />
-        </Route>
-        <Route path="/analysis">
-          {/* <Analysis /> */}
-          <Admin />
-        </Route>
-        {/* <Route path="/profile">
+    <>
+      {storeCtx.tokenExpired && (
+        <Dialog
+          className="exchange"
+          title="Info"
+          block={true}
+          onConfirm={() => {
+            history.replace({
+              pathname: `/signin`,
+            });
+            window.location.reload();
+          }}
+        >
+          <p className="info__text">{t("token_expired")}</p>
+        </Dialog>
+      )}
+      <Suspense fallback={<div></div>}>
+        <Switch>
+          <Route exact path="/">
+            <Home />
+          </Route>
+          <Route path="/markets">
+            <Exchange />
+          </Route>
+          <Route path="/analysis">
+            {/* <Analysis /> */}
+            <Admin />
+          </Route>
+          {/* <Route path="/profile">
         <Profile />
       </Route>
       <Route path="/wallet">
@@ -86,8 +105,9 @@ const Index = () => {
       <Route path="/notfound">
         <Notfound />
       </Route> */}
-      </Switch>
-    </Suspense>
+        </Switch>
+      </Suspense>
+    </>
   );
 };
